Create README.md
This course introduces the theoretical and applied foundations of Bayesian statistical analysis focused on the problems
encountered by social scientists. The Bayesian paradigm is ideally suited to the type of data analysis required of
social scientists because it recognizes the mobility of population parameters, incorporates prior knowledge that
researchers possess, and updates estimates as new data are observed. This course will introduce the basic principles of
Bayesian statistics to students in the social and behavioral sciences, focusing on examples drawn from sociology,
political science, economics, marketing, psychology, public policy, and anthropology. Models will be estimates with
bugs/jags as well as basic R. Therefore course will make exclusive use of software that is in the public domain, yet is
high in quality.  Most of the examples will be drawn from sociology, political science, economics, marketing,
psychology, public policy, and anthropology.